refactor(trusty-search): extract types.rs from indexer/mod.rs — ratchet allowlist (#607 PR-E follow-up)
Move CodeChunk, SearchQuery, SearchMode, SearchStage, CommitTimings,
ParsedBatch, ChunkSnapshot type definitions from indexer/mod.rs into a new
indexer/types.rs (233 lines). indexer/mod.rs drops from 656 to 441 lines
(now under the 500-line cap), so its allowlist entry is removed per the
ratchet rule.

Quality bar: cargo check, clippy --all-targets, all 1050 tests pass, fmt
clean, scripts/check_line_cap.sh 0 violations.
